fix: resolve CI test failures in optimize scope-aware variable export spec
- Use setVariables(elementInstanceKey, vars, local=true) in service task workers
  so variables are created at element-instance scope (scopeKey !== processInstanceKey)
  rather than being promoted to root scope
- Correct TC-23 assertion: propagateAllChildVariables places child vars at root scope
  in the parent process, so scopeKey === processInstanceKey is the expected behaviour
- Type createWorker handler with ZBWorkerTaskHandler to avoid any annotations
- Remove prose comments and align test.describe to parallel convention
