+++ qa/c8-orchestration-cluster-e2e-test-suite/utils/requestHelpers/index.ts
@@ -96,3 +96,14 @@ export {
 } from './global-task-listener-requestHelpers';
 export {type AuditLog} from './audit-log-requestHelpers';
 export {evaluateExpression, EXPRESSION_URL} from './expression-requestHelpers';
+export {
+  type VariableRecord,
+  isRootScope,
+  isLocalScope,
+  getRootScopeVariables,
+  getLocalScopeVariables,
+  getAllProcessInstanceVariables,
+  assertNoVariablesForProcessInstance,
+  getVariablesByName,
+  getVariablesByPatterns,
+} from './optimize-variable-requestHelpers';
