+++ qa/c8-orchestration-cluster-e2e-test-suite/utils/zeebeClient.ts
@@ -6,10 +6,8 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {readFileSync} from 'fs';
-import {basename} from 'path';
 import {Camunda8} from '@camunda8/sdk';
-import {JSONDoc} from '@camunda8/sdk/dist/zeebe/types.js';
+import {JSONDoc, ZBWorkerTaskHandler} from '@camunda8/sdk/dist/zeebe/types.js';
 
 const c8 = new Camunda8({
   CAMUNDA_AUTH_STRATEGY: process.env.CAMUNDA_AUTH_STRATEGY as
@@ -43,28 +41,6 @@ const deploy = async (processFilePaths: string[]) => {
   try {
     const results = await zeebe.deployResourcesFromFiles(processFilePaths);
     return results;
-  } catch (error) {
-    console.error('Deployment failed:', error);
-    throw error;
-  }
-};
-
-const deployWithSubstitutions = async (
-  bpmnFilePath: string,
-  substitutions: Record<string, string>,
-): Promise<void> => {
-  let content = readFileSync(bpmnFilePath, 'utf-8');
-  for (const [placeholder, replacement] of Object.entries(substitutions)) {
-    if (!content.includes(placeholder)) {
-      throw new Error(
-        `Placeholder '${placeholder}' not found in BPMN file '${bpmnFilePath}'`,
-      );
-    }
-    content = content.split(placeholder).join(replacement);
-  }
-  const name = basename(bpmnFilePath);
-  try {
-    await zeebe.deployResources([{content, name}]);
   } catch (error) {
     console.error('Deployment failed:', error);
     throw error;
@@ -117,8 +93,7 @@ const createWorker = (
   taskType: string,
   shouldFail: boolean = false,
   variables: JSONDoc = {},
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  handler?: (job: any) => any,
+  handler?: ZBWorkerTaskHandler,
   timeout?: number,
 ) => {
   return zeebeGrpc.createWorker({
@@ -157,7 +132,6 @@ const setVariables = async (
 
 export {
   deploy,
-  deployWithSubstitutions,
   createInstances,
   generateManyVariables,
   checkUpdateOnVersion,
